+++ internal/httputil/handlers.go
@@ -1,0 +1,18 @@
+package httputil // import "github.com/pomerium/pomerium/internal/httputil"
+
+import (
+	"net/http"
+)
+
+func HealthCheck(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodGet && r.Method != http.MethodHead {
+		http.Error(w, http.StatusText(http.StatusMethodNotAllowed), http.StatusMethodNotAllowed)
+		return
+	}
+	w.Header().Set("Content-Type", "text/plain")
+	w.WriteHeader(http.StatusOK)
+	if r.Method == http.MethodGet {
+		w.Write([]byte(http.StatusText(http.StatusOK)))
+	}
+	return
+}